New classes setup in new `Kirby\File` namespace
