+++ src/File/Asset.php
@@ -1,17 +1,14 @@
 <?php
 
-namespace Kirby\Cms;
-
-use Kirby\Toolkit\Properties;
+namespace Kirby\File;
 
 /**
  * Anything in your public path can be converted
  * to an Asset object to use the same handy file
- * methods and thumbnail generation as for any other
- * Kirby files. Pass a relative path to the Asset
- * object to create the asset.
+ * methods as for any other Kirby files. Pass a
+ * relative path to the class to create the asset.
  *
- * @package   Kirby Cms
+ * @package   Kirby File
  * @author    Bastian Allgeier <bastian@getkirby.com>
  * @link      https://getkirby.com
  * @copyright Bastian Allgeier GmbH
@@ -19,36 +16,28 @@ use Kirby\Toolkit\Properties;
  */
 class Asset
 {
-    use FileFoundation;
-    use FileModifications;
-    use Properties;
+    use HasFile;
 
     /**
+     * Relative file path
+     *
      * @var string
      */
     protected $path;
 
+
     /**
-     * Creates a new Asset object
-     * for the given path.
+     * Creates a new Asset object for the given path.
      *
      * @param string $path
      */
     public function __construct(string $path)
     {
-        $this->setPath(dirname($path));
-        $this->setRoot($this->kirby()->root('index') . '/' . $path);
-        $this->setUrl($this->kirby()->url('index') . '/' . $path);
-    }
-
-    /**
-     * Returns the alternative text for the asset
-     *
-     * @return null
-     */
-    public function alt()
-    {
-        return null;
+        $this->setProperties([
+            'path' => dirname($path),
+            'root' => $this->kirby()->root('index') . '/' . $path,
+            'url'  => $this->kirby()->url('index') . '/' . $path
+        ]);
     }
 
     /**
